add TypeScript integration and prettier config
