+++ projects/sttic/.prettierrc
@@ -1,0 +1,8 @@
+{
+	"arrowParens": "avoid",
+	"printWidth": 120,
+	"semi": true,
+	"singleQuote": true,
+	"trailingComma": "none",
+	"useTabs": true
+}
